Chore: Move SpfBuilder to its own Modile file
Move code to new file and fix resulting errors to return to compiling
status.
