+++ src/spf/tests/builder/ip4.rs
@@ -1,6 +1,6 @@
 mod parse {
 
-    use crate::spf::SpfBuilder;
+    use crate::SpfBuilder;
 
     #[test]
     fn valid() {
